+++ src/test/java/delayOptionsTests/DelayOptions_TypeTest.java
@@ -1,7 +1,7 @@
-package delayOptionsTest;
+package delayOptionsTests;
 
 //TODO: finish this
-public class DelayOptions_ConstructorTest {
+public class DelayOptions_TypeTest {
     /**
      * Main method of this test.
      * Run this to start.
